backend/requirements.txt
Django==5.2.4
djangorestframework
django-cors-headers
pymysql
python-dotenv
Pillow
tensorflow